improved Bookings page
