+++ client/src/pages/Routes.js
@@ -5,7 +5,7 @@ import ProtectedRoute from '../components/Auth/hoc/ProtectedRoute';
 import Home from './Home/Home';
 import Auth from './Auth/Auth';
 import Events from './Events/Events';
-import Bookings from './Bookings';
+import Bookings from './Bookings/Bookings';
 
 import {
   ROOT_PATH,
